BAH-4821 | Refactor. Rename CommonSearch Keys And Locales
