+++ apps/clinical/public/locales/locale_en.json
@@ -2,7 +2,7 @@
   "ABNORMAL_VALUE": "Abnormal Value",
   "ADD_ALLERGY_NOTE": "Add Note",
   "ADD_ALLERGY_NOTE_PLACEHOLDER": "Add notes for the Allergy",
-  "ADVANCED_SEARCH_LABEL": "Advanced Search",
+  "COMMON_SEARCH_LABEL": "Common Search",
   "ALLERGEN": "Allergen",
   "ALLERGIES_ADDED_ALLERGIES": "Added Allergies",
   "ALLERGIES_DISPLAY_CONTROL_HEADING": "Allergies",
