+++ apps/clinical/public/locales/locale_es.json
@@ -2,7 +2,7 @@
   "ABNORMAL_VALUE": "Valor Anormal",
   "ADD_ALLERGY_NOTE": "Agregar Nota",
   "ADD_ALLERGY_NOTE_PLACEHOLDER": "Agregar notas para la alergia",
-  "ADVANCED_SEARCH_LABEL": "Búsqueda Avanzada",
+  "COMMON_SEARCH_LABEL": "Búsqueda Común",
   "ALLERGEN": "Alérgeno",
   "ALLERGIES_ADDED_ALLERGIES": "Alergias Agregadas",
   "ALLERGIES_DISPLAY_CONTROL_HEADING": "Alergias",
