+++ packages/bahmni-widgets/src/search/index.ts
@@ -1,4 +1,4 @@
-import './advancedSearch';
+import './commonSearch';
 
 export {
   registerSearchWidget,
